+++ src/extension.ts
@@ -3,6 +3,7 @@
 import * as vscode from "vscode";
 // import * as l10n from "@vscode/l10n";
 import { registerSetUnicornUnityDirectoryCommand } from "./commands/setUnicornUnityDirectory";
+import { registerCopyThemeIconsFromRemoteCommand } from "./commands/copyThemeIcon";
 import { registerCopyThemeMiniIconsFromRemoteCommand } from "./commands/copyThemeMiniIcon";
 import * as commonUtils from "./utils/commonUtils";
 import { register } from "module";
@@ -55,33 +56,36 @@ export function activate(context: vscode.ExtensionContext) {
 	// update status bar item once at start
 	updateStatusBarItem();
 
-	vscode.languages.registerHoverProvider(
-		"lua",
-		new (class implements vscode.HoverProvider {
-			provideHover(
-				_document: vscode.TextDocument,
-				_position: vscode.Position,
-				_token: vscode.CancellationToken
-			): vscode.ProviderResult<vscode.Hover> {
-				const commentCommandUri = vscode.Uri.parse(
-					`command:editor.action.addCommentLine`
-				);
-				const contents = new vscode.MarkdownString(
-					`[Add comment](${commentCommandUri})`
-				);
+	// comment meaningless hover provider.
+	// vscode.languages.registerHoverProvider(
+	// 	"lua",
+	// 	new (class implements vscode.HoverProvider {
+	// 		provideHover(
+	// 			_document: vscode.TextDocument,
+	// 			_position: vscode.Position,
+	// 			_token: vscode.CancellationToken
+	// 		): vscode.ProviderResult<vscode.Hover> {
+	// 			const commentCommandUri = vscode.Uri.parse(
+	// 				`command:editor.action.addCommentLine`
+	// 			);
+	// 			const contents = new vscode.MarkdownString(
+	// 				`[Add comment](${commentCommandUri})`
+	// 			);
 
-				// To enable command URIs in Markdown content, you must set the `isTrusted` flag.
-				// When creating trusted Markdown string, make sure to properly sanitize all the
-				// input content so that only expected command URIs can be executed
-				contents.isTrusted = true;
+	// 			// To enable command URIs in Markdown content, you must set the `isTrusted` flag.
+	// 			// When creating trusted Markdown string, make sure to properly sanitize all the
+	// 			// input content so that only expected command URIs can be executed
+	// 			contents.isTrusted = true;
 
-				return new vscode.Hover(contents);
-			}
-		})()
-	);
+	// 			return new vscode.Hover(contents);
+	// 		}
+	// 	})()
+	// );
 
 	// Register commands
+	// Predifined command registration
 	registerSetUnicornUnityDirectoryCommand(context);
+	registerCopyThemeIconsFromRemoteCommand(context);
 	registerCopyThemeMiniIconsFromRemoteCommand(context);
 }